requirements.txt
pandas
matplotlib
scipy
python-dotenv